Fix abstraction
