+++ src/Abstractions/NexusMods.Abstractions.EventBus/IEventBusMessage.cs
@@ -1,0 +1,9 @@
+using JetBrains.Annotations;
+
+namespace NexusMods.Abstractions.EventBus;
+
+/// <summary>
+/// Represents a message.
+/// </summary>
+[PublicAPI]
+public interface IEventBusMessage;
