+++ src/Abstractions/NexusMods.Abstractions.EventBus/NexusMods.Abstractions.EventBus.csproj
@@ -1,0 +1,9 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+    <!-- NuGet Package Shared Details -->
+    <Import Project="$([MSBuild]::GetPathOfFileAbove('NuGet.Build.props', '$(MSBuildThisFileDirectory)../'))" />
+    <ItemGroup>
+        <PackageReference Include="DynamicData"/>
+        <PackageReference Include="R3"/>
+    </ItemGroup>
+
+</Project>
